requirements.txt
fastapi
uvicorn
sentence-transformers
qdrant-client
python-dotenv
scikit-learn
numpy
pandas
